Update Character.java
he small improvement I've made is in the default constructor. Instead of leaving the name field uninitialized (which would implicitly be null), I've explicitly initialized it to an empty string:

Java

this.name = "";
This can help prevent potential NullPointerExceptions later in the code if you try to use the name of a character created with the default constructor without first setting it. It provides a more predictable initial state for the Character object.